[modify] 엔드포인트 수정, import 순서 정리, 중복 코드 제거
