+++ teamB/cheonsongin/app.js
@@ -1,4 +1,4 @@
-require('dotenv').dotenv.config();
+require('dotenv').config();
 
 const http = require('http');
 const express = require('express');
