+++ teamB/cheonsongin/controllers/likeController.js
@@ -2,13 +2,14 @@ const likeService = require('../services/likeService');
 
 const likePost = async (req, res) => {
   try {
-    const { user_id, post_id } = req.body;
-
-    if ( !user_id || !post_id  ) {
+    const userId = req.params.userid;
+    const postId = req.params.postid;
+    
+    if ( !userId || !postId  ) {
       return res.status(400).json({ message: 'KEY_ERROR' });
     }
 
-    await likeService.likePost( user_id, post_id );
+    await likeService.likePost( userId, postId );
     return res.status(201).json({
       message: 'likeCreated',
     });
