+++ teamB/cheonsongin/models/likeDao.js
@@ -1,32 +1,11 @@
-const { DataSource } = require('typeorm');
-
-const appDataSource = new DataSource({
-  type: process.env.TYPEORM_CONNECTION,
-  host: process.env.TYPEORM_HOST,
-  port: process.env.TYPEORM_PORT,
-  username: process.env.TYPEORM_USERNAME,
-  password: process.env.TYPEORM_PASSWORD,
-  database: process.env.TYPEORM_DATABASE
-});
-
-appDataSource.initialize()
-  .then(() => {
-    console.log('Data Source has been initialized!');
-  })
-  .catch((err) => {
-    console.error('Error occurred during Data Source initialization', err);
-    appDataSource.destroy();
-  });
-
-const postLike = async ( user_id, post_id ) => {
+const postLike = async ( userId, postId ) => {
   try {
     return await appDataSource.query(
       `INSERT INTO likes(
         user_id,
         post_id
-      ) VALUES (?, ?);
+      ) VALUES (${userId}, ${postId});
       `,
-      [ user_id, post_id ]
     );
   } catch (err) {
     const error = new Error('INVALID_DATA_INPUT');
