+++ teamB/cheonsongin/models/userDao.js
@@ -1,23 +1,3 @@
-const { DataSource } = require('typeorm');
-
-const appDataSource = new DataSource({
-  type: process.env.TYPEORM_CONNECTION,
-  host: process.env.TYPEORM_HOST,
-  port: process.env.TYPEORM_PORT,
-  username: process.env.TYPEORM_USERNAME,
-  password: process.env.TYPEORM_PASSWORD,
-  database: process.env.TYPEORM_DATABASE
-});
-
-appDataSource.initialize()
-  .then(() => {
-    console.log('Data Source has been initialized!');
-  })
-  .catch((err) => {
-    console.error('Error occurred during Data Source initialization', err);
-    appDataSource.destroy();
-  });
-
 const createUser = async ( name, email, password, profileImage ) => {
   try {
     return await appDataSource.query(
@@ -37,6 +17,33 @@ const createUser = async ( name, email, password, profileImage ) => {
   }
 };
 
+const selectUserPost = async (userId) => {
+  try {
+    const users = await appDataSource.query(
+      `SELECT
+        u.id AS userId,
+        u.profile_image AS userProfileImage
+      FROM users u
+      WHERE id=${userId}`
+    );
+    const postsings = await appDataSource.query(
+      `SELECT
+        p.id AS postingId,
+        p.post_image AS postingImageUrl,
+        p.content AS postingContent
+      FROM posts p
+      WHERE user_id=${userId}`
+    );
+    users[0].posting = postsings;
+    return users;
+  } catch (err) {
+    const error = new Error('INVALID_DATA_INPUT');
+    error.statusCode = 500;
+    throw error;
+  }
+};
+
 module.exports = {
-  createUser
+  createUser,
+  selectUserPost
 }
