+++ teamB/cheonsongin/routes/userRouter.js
@@ -1,9 +1,11 @@
 const express = require('express');
-const userController = require('../controllers/userController');
-
 const router = express.Router();
 
+const userController = require('../controllers/userController');
+
 router.post('/signup', userController.signUp);
+router.get('/:userid', userController.viewUserId);
+
 
 module.exports = {
   router
