+++ teamB/cheonsongin/services/userService.js
@@ -1,10 +1,10 @@
 const userDao = require('../models/userDao');
 
 const signUp = async (name, email, password, profileImage) => {
-  const pwValidation = new RegExp(
+  const PASSWORD_REGEX = new RegExp(
     '^(?=.*[A-Za-z])(?=.*[0-9])(?=.*[!@#$%^&*])(?=.{8,20})'
   );
-  if (!pwValidation.test(password)) {
+  if (!PASSWORD_REGEX.test(password)) {
     const err = new Error('PASSWORD_IS_NOT_VALID');
     err.statusCode = 409;
     throw err;
@@ -19,6 +19,12 @@ const signUp = async (name, email, password, profileImage) => {
   return createUser;
 };
 
+const viewUserId = async (userId) => {
+  const selectUserPost = await userDao.selectUserPost(userId);
+  return selectUserPost;
+}
+
 module.exports = {
-  signUp
+  signUp,
+  viewUserId
 }
